add last updater

+++ app/Http/Controllers/Api/RepositoriesController.php
@@ -10,60 +10,55 @@ use App\Http\Requests\Api\RepositoryRequest;
 
 class RepositoriesController extends Controller
 {
-    public function store(RepositoryRequest $request, Repository $repository)
+    public function store(Repository $repository, RepositoryRequest $request)
     {
-        $repository->fill($request->all());
+        $attributes = $request->only(['name']);
+        $repository->fill($attributes);
         $repository->user_id = $this->user()->id;
+        $repository->last_updater_id = $this->user()->id;
         $repository->save();
 
         return $this->response->item($repository, new RepositoryTransformer())
             ->setStatusCode(201);
     }
 
-    public function update(RepositoryRequest $request, Repository $repository)
+    public function update(Repository $repository, RepositoryRequest $request)
     {
-        // $this->authorize('update', $repository);
-
-        $repository->update($request->all());
+        $this->authorize('update', $repository);
+        $attributes = $request->only(['name']);
+        $repository->update($attributes);
+        $repository->last_updater_id = $this->user()->id;
+        $repository->save();
         return $this->response->item($repository, new RepositoryTransformer());
     }
 
     public function destroy(Repository $repository)
     {
-        // $this->authorize('destroy', $repository);
-
+        $this->authorize('destroy', $repository);
         $repository->delete();
         return $this->response->noContent();
     }
 
-    public function index(RepositoryRequest $request, Repository $repository)
+    public function show(Repository $repository)
     {
-        $query = $repository->query();
-
-        // if ($categoryId = $request->category_id) {
-        //     $query->where('category_id', $categoryId);
-        // }
-
-        switch ($request->order) {
-            case 'recent':
-                $query->recent();
-                break;
-
-            default:
-                $query->recentUpdated();
-                break;
-        }
-
-        $repositories = $query->paginate(20);
-
-        return $this->response->paginator($repositories, new RepositoryTransformer());
+        $this->authorize('show', $repository);
+        return $this->response->item($repository,new RepositoryTransformer());
     }
 
-    public function userIndex(User $user, RepositoryRequest $request)
+    public function userIndex(User $user)
     {
         $repositories = $user->repositories()->recent()
             ->paginate(20);
 
         return $this->response->paginator($repositories, new RepositoryTransformer());
     }
+
+    public function parterIndex(User $user)
+    {
+        $repositories = $user->parterRepositories()->recent()
+            ->paginate(20);
+
+        return $this->response->paginator($repositories, new RepositoryTransformer());
+    }
+
 }
